save recent writes and display them as colours

+++ Source/SoundGen.cpp
@@ -990,7 +990,7 @@ void CSoundGen::AddCycles(int Count)
 	m_pAPU->AddTime(Count);
 }
 
-uint8 CSoundGen::GetReg(int Chip, int Reg) const
+uint16 CSoundGen::GetReg(int Chip, int Reg) const		// // //
 { 
 	return m_pAPU->GetReg(Chip, Reg);
 }
